Add files via upload
What is TSCA??

TSCA or TSCA will Help to Customise Your Termux in such a way that you will Enjoy using Termux it will give you a morden look and we also have add a feature of password,so now you can also set Password on your Termux and protect it so no one can login in your termux in absence of you. one more feature that we have added is that you can change dollar sign and put anything you want in that place. Hope you guys like this Project,then dont forget to give it a like.
Features of T-Remix

    Fully customise your Termux app.
    Encrypt your Termux app.
    Password protection.
    Welcome voice-message.
    Change the $ (DOLLAR) sign with a modern look.
    New banner.
    Login banner is for HACKERs only.
    Show off to other with hacker attitude.